refactor last updated and navigation
