+++ src/App.js
@@ -4,7 +4,7 @@ import 'react-toastify/dist/ReactToastify.css';
 import './App.css';
 import Comments from './components/Comment/Comments';
 import CreateComment from './components/Comment/CreateComment';
-import Contact from './components/Contact/Contact';
+import Contact from './components/About/About';
 import Login from './components/Login/Login';
 import Navigation from './components/Navigation/Navigation';
 import CreatePost from './components/Post/CreatePost';
@@ -21,7 +21,7 @@ function App() {
         <Navigation />
         <Routes>
           <Route path="/" element={<Posts />} />
-          <Route path="/contacts" element={<Contact />} />
+          <Route path="/about" element={<Contact />} />
           <Route path="/myposts" element={<MyPosts />} />
           <Route path="/user" element={<User />} />
           <Route path="/signup" element={<Signup />} />
