+++ src/components/Post/Posts.js
@@ -4,13 +4,10 @@ import { NavLink } from 'react-router-dom';
 import { getPosts } from '../../redux/actions/Post';
 import LikeCommentCount from '../LikeCommentCount/LikeCommentCount';
 import User from '../User';
+import LatestPost from './LatestPost';
 
 const Posts = () => {
   const posts = useSelector((state) => state.PostReducer);
-  const ids = posts.posts.map((post) => post.post_id);
-  const maxId = Math.max(...ids);
-  console.log('max Id: ', maxId);
-  const latestPost = posts.posts.filter((post) => post.post_id === maxId);
   const dispatch = useDispatch();
 
   useEffect(() => {
@@ -19,32 +16,11 @@ const Posts = () => {
 
   const setPostId = (id) => {
     localStorage.setItem('postid', id);
-    // navigate(`/post/${postId}`);
   };
   return (
     <>
       <div className="flex flex-col m-2">
-        <h2 className="ml-6 text-3xl font-bold">Latest Post</h2>
-        {latestPost?.map((post) => (
-          <div className="flex flex-col md:flex-row ml-6 mt-2 pb-4 space-x-12 border-b-2 md:shrink-0" key={maxId}>
-            <img className="rounded object-cover h-56" src={post.image} alt="img" />
-            <div className="flex flex-col justify-center border-b-3">
-              <h4 className="font-medium text-lg text-4xl">
-                {post.title}
-              </h4>
-              <div className="flex">
-                <p className="w-60 truncate">{post.text}</p>
-                <NavLink to="/post"><button className="text-green-500" type="button" onClick={() => setPostId(post.post_id)}>more</button></NavLink>
-              </div>
-              <LikeCommentCount id={post.post_id} />
-              <div>
-                Author:
-                {' '}
-                <div className="inline text-red-500"><User id={post.user_id} /></div>
-              </div>
-            </div>
-          </div>
-        ))}
+        <LatestPost />
         <div className="flex flex-wrap m-4">
           {posts.posts?.map((post) => (
             <div className="flex flex-col items-left p-2 space-y-4 border m-2" key={post.post_id}>
